+++ RapidWebView/src/main/java/in/workindia/rapidwebview/constants/BroadcastConstants.kt
@@ -2,6 +2,8 @@ package `in`.workindia.rapidwebview.constants
 
 class BroadcastConstants {
     companion object {
+        const val UPLOADED_URL = "uploadedUrl"
+        const val UPLOADED_FILE_NAME = "uploadedFileName"
         const val NATIVE_CALLBACK_ACTION = "nativeActionListener"
         const val UPLOAD = "upload"
         const val PERMISSION = "permission"