+++ samples/RapidWebViewSdkDocsSite/pages/docs/js-interface/functions.tsx
@@ -90,8 +90,8 @@ const FUNCTION_DOCS = [
   {
     id: 13,
     title: "Request permission",
-    description: 'To ask for android permission. It requires 3 parameters, an array containing all the permissions needed, rationale Text(text to show on permission dialog), and a callback. A callback is a reverse bridge between the Web and Android, it acts as a broadcast receiver from android to your website, it can be used to know whether your permission resulted in a success or a failure. If this information is not needed it can be passed as null.',
-    code: '.requestPermissions(["android.permission.CALL_PHONE"], "Permission required",null)',
+    description: 'To ask for android permission. It requires 2 parameters, an array containing all the permissions needed, rationale Text(text to show on permission dialog), and a callback. Callbacks from native can be used by using javascript event listener with the event "rapid-web-view-permission-listener" which will return object { detail: { "status" : "success","uploadUrl" : "$uploadUrl","uploadFileName" : "$fileName"} }',
+    code: '.requestPermissions(["android.permission.CALL_PHONE"], "Permission required")',
   },
   {
     id: 14,
@@ -108,8 +108,8 @@ const FUNCTION_DOCS = [
   {
     id: 16,
     title: "Open native file upload interface",
-    description: 'To upload a file using native file upload interface. It accepts 3 parameters, the file type, upload Url(where the file will be uploaded), and third a callback, this callback is either a javascript function or url to open once file upload activity is finished. Callback function can receive whether upload file succeeded by embedding params in the callback string. ',
-    code: '.uploadFile("doc", "www.example.com/uploadfile", "https://example.com/api/fileUploaded/?data=$params")',
+    description: 'To upload a file using native file upload interface. It accepts 3 parameters, the file type, upload Url(where the file will be uploaded), and third a method to either "PUT" or "POST", Callbacks from native can be used by using javascript event listener with the event "rapid-web-view-upload-listener" which will return object { detail: { "status" : "success","uploadUrl" : "$uploadUrl","uploadFileName" : "$fileName"} }',
+    code: '.uploadFile("doc", "www.example.com/uploadfile", "PUT|POST")',
   },
 ];